+++ geode-core/src/main/java/org/apache/geode/internal/admin/statalerts/FunctionHelper.java
@@ -143,7 +143,7 @@ public class FunctionHelper {
       for (final Number val : vals) {
         sum = sum + val.doubleValue();
       }
-      return Double.valueOf(sum);
+      return sum;
     } catch (VirtualMachineError err) {
       SystemFailure.initiateFailure(err);
       // If this ever returns, rethrow the error. We're poisoned
@@ -167,7 +167,7 @@ public class FunctionHelper {
    */
   public static Number AVG(Number[] vals) {
     try {
-      return Double.valueOf(SUM(vals).doubleValue() / vals.length);
+      return SUM(vals).doubleValue() / vals.length;
     } catch (VirtualMachineError err) {
       SystemFailure.initiateFailure(err);
       // If this ever returns, rethrow the error. We're poisoned
